refactor(skeleton): fix container & config
