+++ main.py
@@ -3,7 +3,7 @@ import os
 import click
 import uvicorn
 
-from application.core.container.config import config
+from application.core.config.config_container import config
 
 
 @click.command()
@@ -23,7 +23,7 @@ def main(env: str, debug: bool):
     print(f"RUNNING ENV is {env}")
     os.environ["DEBUG"] = str(debug)
     uvicorn.run(
-        app="application.server:domain",
+        app="application.server:app",
         host=config.APP_HOST(),
         port=config.APP_PORT(),
         reload=True if config.ENV() != "production" else False,
